ml/requirements.txt
# ML/AI Platform Requirements

## Core ML Libraries
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.2
lightgbm==4.1.0
xgboost==2.0.2
numpy==1.24.3
pandas==2.1.3

## MLOps
mlflow==2.9.0
dvc==3.30.0
dvc-s3==3.0.1
evidently==0.4.12

## Serving
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

## NLP & Embeddings
sentence-transformers==2.2.2
openai==1.3.0
tiktoken==0.5.1

## Data Processing
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
spacy==3.7.2

## Explainability
shap==0.43.0

## Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

## Vector DB
pymilvus==2.3.3
pinecone-client==2.2.4

## Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
click==8.1.7
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0

## Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

## Development
jupyter==1.0.0
ipykernel==6.27.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
